Move getThirdPartyMessage to api/backend
